feat(1-3385): more cleanup
