+++ frontend/src/component/common/DialogFormTemplate/ConfigButtons/ConfigButton.styles.tsx
@@ -1,16 +1,4 @@
-import { Popover, styled } from '@mui/material';
-
-export const StyledPopover = styled(Popover)(({ theme }) => ({
-    '& .MuiPaper-root': {
-        borderRadius: `${theme.shape.borderRadiusMedium}px`,
-        paddingInline: 0,
-        paddingTop: theme.spacing(1.5),
-        display: 'flex',
-        flexDirection: 'column',
-        gap: theme.spacing(1),
-        maxHeight: '70vh',
-    },
-}));
+import { styled } from '@mui/material';
 
 export const ButtonLabel = styled('span', {
     shouldForwardProp: (prop) => prop !== 'labelWidth',
