+++ frontend/src/component/common/DialogFormTemplate/ConfigButtons/ConfigButton.tsx
@@ -1,13 +1,10 @@
 import { v4 as uuidv4 } from 'uuid';
 import { type FC, type ReactNode, useRef, type PropsWithChildren } from 'react';
 import { Box, Button } from '@mui/material';
-import {
-    StyledPopover,
-    ButtonLabel,
-    StyledTooltipContent,
-} from './ConfigButton.styles';
+import { ButtonLabel, StyledTooltipContent } from './ConfigButton.styles';
 import { TooltipResolver } from 'component/common/TooltipResolver/TooltipResolver';
 import { ScreenReaderOnly } from 'component/common/ScreenReaderOnly/ScreenReaderOnly';
+import { StyledPopover } from './shared.styles';
 
 export type ConfigButtonProps = {
     button: {
